third_party/blink/web_tests/external/wpt/css/cssom/cssom-fontfacerule-constructors-expected.txt

This is a testharness.js-based test.
[FAIL] The style attribute must return a CSSFontFaceDescriptors block
  assert_equals: expected "[object CSSFontFaceDescriptors]" but got "[object CSSStyleDeclaration]"
Harness: the test ran to completion.
